refactor: rename useBackgroundCanvas import path for consistency
