+++ src/features/editor/hooks/use-background-canvas.tsx
